visualizer/requirements.txt
attrs==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
cykhash==2.0.1
Cython==0.29.34
Fiona==1.9.3
Flask==2.2.3
geographiclib==2.0
geopandas==0.12.2
geopy==2.3.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
munch==2.5.0
numpy==1.24.3
packaging==23.1
paho-mqtt==1.6.1
pandas==2.0.1
pygeos==0.14
pyproj==3.5.0
pyrobuf==0.9.3
pyrosm==0.6.1
python-dateutil==2.8.2
python-rapidjson==1.10
pytz==2023.3
requests==2.29.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
threaded==4.1.0
tzdata==2023.3
urllib3==1.26.15
Werkzeug==2.2.3
